xterm-safe colors, Spell[xxx]
The light :terminal palette is now xterm-safe.
It works well enough in all contexts I've tested it in, and it gives
users who might want to add features to quiet a reference palette to
work with.
The various Spell[xxx] groups should now work properly in 8/16c.

Had to rejig the template quite a bit, but it was probably worth it.