First stab at a nord/iceberg clone.
Quick and dirty hack on top of `quiet`. The UI needs to be thoroughly
checked for collisions with the main palette.
The palette was shamelessly lifted from nord and iceberg, then adjusted
to be unambiguous in a 256c context, assuming a pure black, or dark grey
background.
The variants are mainly intended for quick reference between ideas
during the PR review process, and I expect most, if not all of them,
to be culled before release.