habamax: no color Title
Title is used to highlight number of split windows inside a TabPage.

When colored, it might clash with TabLineSel background making number of
windows unreadable.

Until a separate highlight group is implemented in Vim, it is safer to
have no-color Title unless one can find a combination of fg/bg colors
that would work OK for both Title and TabLine(Sel).