Remove PopupSelected highlight links
After vim/vim#16676, Vim is guaranteed to use the PmenuSel group if
PopupSelected is not defined in a color scheme. Because of that, there
is no need to manually create a highlight group for PopupSelected,
unless the color scheme specifically intends to use a different color,
which none of the color schemes in this repository does.