Add the ability to take screen dumps of the color schemes
This commit adds functions to take screen dumps of the
output of the `sample*.vim` scripts already in the repo.

This enables the automatic comparison of the visual
appearance of a color scheme before and after some
changes are applied. In particular, this may turn useful
to avoid unwanted changes when one update a template
in a way that should not affect the visual appearance of a
color scheme.