Make the screen output of sample_windows.vim more predictable.
The command-line completion would change if files or directories
were created. Adding `sample` ensures that the command line
shows the same items even if the content of the directory changes
between a screen dump and the next one (well, at least if files
whose name starts with `sample` are created).