Fix some visual discrepancies with the color schemes currently in Vim.
- Added `Float`, `Function`, `Number` and `QuickFixLine to `_tco_zero`.
   They are all defined to generate `term=NONE`. Accordingly, default
   definitions for those highlight groups have been added to the color
   schemes in which they were not present.

I consider the remaining (few) visual differences wrt the color schemes
before this commit as  *fixes of shortcomings in master.* Specifically:

- `lunaperche` did not define `String` for dark/light 8-color terminals.
- `lunaperche` did not define `Visual` for 8/16-color terminals (except
   for the combination 16/light), falling back to the `default` color scheme.
- `retrobox`'s template removes `bold` for `Error` in 8/16-color terminals
   (both backgrounds), but that definition was not applied to the color scheme.
   This is fixed in this PR
- `wildcharm` (dark) had this definition:

```
Variant: gui 256
EndOfBuffer                    colorNonT      none
Variant: gui 256 16
EndOfBuffer                    color08        none
```

So, it is not clear what the foreground should be for `gui` and `256`.
By comparing with the light background and considering what the intent
could have been here, I've opted for `colorNonT` by default, and `color08`
for 16-color terminals. This results in a different `EndOfBuffer` background
in dark wildcharm than what is currently in Vim.

Finally, for wildcharm, the version in Vim's runtime has a `NONE` background
for `VertSplit`, but the version in this repo has a solid `VertSplit`. I assume that this
is a change that has not been pushed to Vim yet. I've kept the solid appearance.

Signed-off-by: Lifepillar <lifepillar@lifepillar.me>